Remove direct flake usage of `nixpkgs` input (#16)
The `nixpkgs` dependency was only being used for library functions, and
these functions are only needed within `garnixModules.default` which
receives `lib` from `garnix-lib`.

However, this change does *not* remove the `nixpkgs` input since it is
still being used as an input to `dream2nix`. This makes it clear that
`dream2nix` is consuming a `nixpkgs` and makes it easy to override.